refactor: Introduce AbstractPactBuilder for shared verify functionality
