+++ src/PactNet/Drivers/ICompletedPactDriver.cs
@@ -14,5 +14,16 @@ namespace PactNet.Drivers
         /// <returns>Status code</returns>
         /// <exception cref="InvalidOperationException">Failed to write pact file</exception>
         void WritePactFile(string directory);
+
+        /// <summary>
+        /// Create the mock server for the current pact
+        /// </summary>
+        /// <param name="host">Host for the mock server</param>
+        /// <param name="port">Port for the mock server, or null to allocate one automatically</param>
+        /// <param name="tls">Enable TLS</param>
+        /// <param name="transport">The transport to use (i.e. http, https, grpc). Must be a valid UTF-8 NULL-terminated string, or NULL or empty, in which case http will be used.</param>
+        /// <returns>Mock server port</returns>
+        /// <exception cref="InvalidOperationException">Failed to start mock server</exception>
+        IMockServerDriver CreateMockServer(string host, int? port, bool tls, string transport = null);
     }
 }
