+++ src/PactNet/Drivers/Message/MessagePactDriver.cs
@@ -7,15 +7,12 @@ namespace PactNet.Drivers.Message
     /// </summary>
     internal class MessagePactDriver : AbstractPactDriver, IMessagePactDriver
     {
-        private readonly PactHandle pact;
-
         /// <summary>
         /// Initialises a new instance of the <see cref="MessagePactDriver"/> class.
         /// </summary>
         /// <param name="pact">Pact handle</param>
         internal MessagePactDriver(PactHandle pact) : base(pact)
         {
-            this.pact = pact;
         }
 
         /// <summary>
