+++ tests/PactNet.Tests/Drivers/FfiIntegrationTests.cs
@@ -54,7 +54,7 @@ namespace PactNet.Tests.Drivers
                 interaction.WithResponseHeader("X-Response-Header", "value2", 1);
                 interaction.WithResponseBody("application/json", @"{""foo"":42}");
 
-                using IMockServerDriver mockServer = pact.CreateMockServer("127.0.0.1", null, false);
+                using IMockServerDriver mockServer = pact.CreateMockServer("127.0.0.1", null, false, "http");
 
                 var client = new HttpClient { BaseAddress = mockServer.Uri };
                 client.DefaultRequestHeaders.Add("X-Request-Header", new[] { "request1", "request2" });
